moves `getWebhook` to `service/db`
