+++ src/service/db/getWebhook.ts
@@ -1,4 +1,4 @@
-import prisma from '../client';
+import prisma from '../../client';
 
 export async function getWebhookByWebhookID(webhookID: string) {
   try {
